Extract checked assertions
